+++ bin/create.js
@@ -1,25 +1,25 @@
 #!/usr/bin/env node
 
+import command from "@m4rch/command"
 import { existsSync, mkdirSync, readdirSync } from "fs"
 import { join } from "path"
-import { error } from "../src/functions.js"
+import { error } from "../src/utils.js"
 
-import { default as Command } from "../src/command.js"
-import { default as index } from "../src/index.js"
+import options from "./options.js"
+import index from "../src/index.js"
 
 let _argv = process.argv.slice(2)
 let rootPath = process.cwd()
 
 if (_argv.length) rootPath = join(rootPath, _argv[0])
 if (!existsSync(rootPath) || !readdirSync(rootPath).length) {
-	let command = new Command
-
-	command
-		.get()
-		.then(( answers ) => {
+	command()
+		.get(options)
+		.action(( answers ) => {
 			if (!existsSync(rootPath)) mkdirSync(rootPath, { recursive: true })
 			index(answers, rootPath)
 		})
+		.run()
 } else {
 	error("can't initiate in a non-empty folder.")
 }