+++ src/utils.js
@@ -1,5 +1,5 @@
-import { exec } from "child_process"
 import { mkdirSync, writeFileSync } from "fs"
+import { exec } from "child_process"
 import { join } from "path"
 
 export async function username () {
@@ -16,9 +16,9 @@ export async function username () {
 	})
 }
 
-export function install ( dir ) {
+export function install ( cwd ) {
 	return new Promise(( resolve, reject ) => {
-		exec("npm install", { cwd: dir }, ( err, stdout, stderr ) => {
+		exec("npm install", { cwd }, ( err, _stdout, stderr ) => {
 			if (!stderr && !err) {
 				resolve()
 			} else {
@@ -30,7 +30,6 @@ export function install ( dir ) {
 
 export function writeFiles ( files, rootPath ) {
 	files.forEach(( file ) => {
-
 		if (file.content != null) {
 			let content = typeof file.content == "object" ? JSON.stringify(file.content, null, "\t") + "\n" : file.content.toString()
 			writeFileSync(join(rootPath, file.name), content)
@@ -38,22 +37,21 @@ export function writeFiles ( files, rootPath ) {
 			mkdirSync(join(rootPath, file.name), { recursive: true })
 			writeFiles(file.files, join(rootPath, file.name))
 		}
-
 	})
 }
 
 export function status ( text ) {
-	console.log(`m4\x1b[36m!notice\x1b[0m ${text}`)
+	console.log(`m4\x1b[36m!notice\x1b[39m ${text}`)
 }
 
 export function success ( text ) {
-	console.log(`m4\x1b[32m!success\x1b[0m ${text}`)
+	console.log(`m4\x1b[32m!success\x1b[39m ${text}`)
 }
 
 export function warn ( text ) {
-	console.log(`m4\x1b[33m!warn\x1b[0m ${text}`)
+	console.log(`m4\x1b[33m!warn\x1b[39m ${text}`)
 }
 
 export function error ( text ) {
-	console.log(`m4\x1b[31m!error\x1b[0m ${text}`)
+	console.log(`m4\x1b[31m!error\x1b[39m ${text}`)
 }